requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.8
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
pymupdf==1.23.7
pytesseract==0.3.10
pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.1
